pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.8",
    "fastapi-admin>=1.0.4",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "jinja2>=3.1.5",
    "jose>=1.0.0",
    "oauthlib>=3.2.2",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.1",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
    "redis>=5.2.1",
    "transitions>=0.9.2",
    "pyyaml>=6.0.2",
    "httpx>=0.28.1",
]
